Charger is enabled with low-level on digital OUT pin of the EV-CC-AC1-M3-CBC-SER Board. A relay is necessary, to physically set High-level (12Volt) to the ENABLE pin, when OUT goes low. Coil of the relay is between OUT and GND. NC-contact between 12V and ENABLE.
The charge controller is great, but has one big disadvantage for my use-case.
When configuring register 4000 for remote enabling via RS485, the default state after power on is "disabled" (which probably makes sense for public chargers).
Default state "charging enabled" after power-up of the EV-CC-AC1-M3 is much better because:

  -  Using as portable charger is possible (the Raspberry is not inside the charger)
  -  When the EVCC system is not working for some reason, charging is still possible.
  -  In the winter, when solar power is barely enaugh for base load, the Raspberry can be shut down, to save power.

The relevant registers in this configuration are:
20000 -> read enabled state
23000 -> disable charging

This registers have to be in factory default:
4000 -> 1
5500 -> 0

The solution was discussed in:
https://github.com/andig/evcc/issues/452