Add scale to SMA meter for usage as PV meter
The SMA meters are and can be used as grid meter and PV meter. EVCC expects PV power to be positiv, but the meter would return it negativ as it is in export direction.

This is solved by adding an optional Scale property that is only used for the Power value. Default value for Scale is `1`