Fix: E3DC external power source now addition to the main power
A condition is used in the render field to retrieve either grid power, photovoltaic power, or battery power, depending on the intended use of the E3DC system. When purpose is "pv", two power values have been subtracted instead of added. To change this, the add field in the power section has been edited to invert the value of 40075 before adding it to the other value, turning the subtraction into an addition.