Ignore a max number of meter no data errors
When an EV is connected, measurements are not available right away, as the initial communication is not done. And due to bugs in the Elli software, the EEBUS stack can not generally wait for data being available and only then set the EV as being connected.

This change ignores a maximum amount of specific EEBUS `no data available` errors, and only propagates them if the error occured more often.

The error will come anyway, if evcc is startet while an EV is connected to the Elli wallbox!