Do not overwrite maxCurrent
The charger should not overwrite the max current with data provided by the charger and instead use what evcc defines.

The max current could in theory change, but practically stays the same, in PMC and ISO connection modes.

With this change it is also possible to use the API to change the loadpoints max current again and define a lower max current than what the charger allows.