Enhance #11990 to use the `BackupBuffer` instead of switching operating mode
**Rationale:**
With #11990 the SonnenBatterie will not be drained when loading vehicles
anymore (thanks @rivengh !!). However, the current solutions forces a switch
of operating modes of the battery between "1" (Manual) and "2" (Auto). This
will pose a problem for users who operate their SonnenBatterie in mode "10"
(Time of Use - ToU) for example to charge the battery when energy is cheap.

To circumvent this problem my solution uses the `BackupBuffer` functionality
of the SonnenBatterie instead. Using the (mislabeled) setting `"EM_USOC"`
one can define the minimum charge the SonnenBatterie will hold back when
discharging.

Using this settings allows for controlled draining of the SonnenBatterie
even during active load plans of evcc.

The current implementation introduces two new config variables:
- `backupbuffer`: The percentage of charge the Sonnenbatterie should
  reserve during normal operation
- `chargingbuffer`: The percentage of charge the Sonnenbatterie should
  hold back during charging operations by evcc

**Note:**
It would be nice if instead of fixed config values we could use the current
setting of the SonnenBatterie for the `BackupBuffer` as determined on
start-up of evcc. Also it would be cool to be able to use the setting
`prioritySoc` as stored in the database for the reserved buffer for charging
times. I don't know whether it's possible to access those in the template
so if it's possible a pointer on how to implement this would be great.

Signed-off-by: Stefan Rubner <stefan@whocares.de>