Use different template paths depending on SB operating mode
This adds yet another config variable `usingtou`. When set to `true`
the EM_USOC type of drain prevention is used. Otherwise the mode
switching variant from @rivengh is used.

This will allow for users to configure the SonnenBatterie to their
current operating mode but has the problem that once the user
changes their decision they'll have to reconfigure evcc as well.

Signed-off-by: Stefan Rubner <stefan@whocares.de>