Huawei Smartlogger: add battery control similar to dongle configuration (#12877)
This is v2 of the configuration. The main difference is that it is
mostly backward compatible to the old smartlogger configuration. The
only change is that it is now using register 40521 (active power without
battery discharge) instead of 40525 (active power including battery
discharge).

The grid information comes from the powermeter which has usually the
unit id 11. The pv informations come directly from smartlogger (unit id
0, hardcoded). The battery information is fetched from the hybrid inverter.

A typical configuration should look like this:

```
meters:
- name: grid
  type: template
  template: huawei-smartlogger
  usage: grid
  id: 11
  host: host.fqdn
  port: 502

- name: pv
  type: template
  template: huawei-smartlogger
  usage: pv
  host: host.fqdn
  port: 502

- name: battery
  type: template
  template: huawei-smartlogger
  usage: battery
  id: 3
  host: host.fqdn
  port: 502
  capacity: 10
```