Add Weidmüller EM110-RTU-2P and EM111-RTU-2P energy meters
The Weidmüller EM110-RTU-2P and EM111-RTU-2P energy meters seem to be
rebranded Eastron SDM230 energy meters. Weidmüller is a popular brand
for electrical installation material in Germany. So it seems much easier
to convince a German eletrician to install a Weidmüller device with a
German brand name and website than an Eastron device with a Chinese
brand name and English website. Therefore, it seems to be a good idea to
document that these Weidmüller devices are supported.