EEbus: upgrade libraries (#13998)
This updates the eebus charger implementation to use the latest version of cemd, eebus, spine and ship.
Only use OSCEV if the a wallbox with ISO15118-2 VW VAS is connected, and the EV is using this. This also enables full charging control and putting the EV to sleep. All other current implementations of the usecase are basically wrong and are no different than using OPEV.