Docs: Update CONTRIBUTING.md
This will add a little bit more information on how to add missing translations to Weblate if there are any. Main thing I learned when finding missing translation was that Github Action will check if string exists in de.toml. This is now added as a note and attached are also 2 example PR where missing translations were added for extra help.