Update EEBUS SHIP library to improve reconnects (#15162)
If a websocket connection was closed, e.g. due to i/o timeout error or remotely denied connection attempt, then a reconnect was only initiated if the remote service announced its mDNS entry again and the code relied on the remote service to initiate the connection. Especially the Elli Connect/Pro devices do not do that, and this situation can appear if the device is connected via WLAN.

This update now uses to currently known mDNS entries upon a disconnect to initiate a connection itself.