fix sunspec-hybrid
The template assumed an inverter with exactly one actual MPPT tracker, and where the second MPPT tracker connected to the battery. It did not use the actual sunspec battery model.

After the change the battery model is used for reading Soc and powers. Active battery control is also added based on the 802 model.