PSA: make climater robust agains missing fields in json
Fix https://github.com/evcc-io/evcc/discussions/19142

Change logic to not assume that climater is running when the PSA API response does not contain the field. `Preconditionning.AirConditioning.Status`. Instead, only assume it is active when the status is `enabled`.
Maybe the API changed, failed or it depends on booked features.