Update doc: Sofar Solar G3 connection
A lot has changed in the last month and there are various methods to connect the inverter where each has several pros and cons. The full doc does not fit in a single line. As my proposal for extended docs on the evcc GitHub was not answerde in #21007, I am proposing to add an external link (that I am regularly updating).