viessmann: remove 'target temperature' parameter
Many devices don't support setting the target temperature, leading to
an error when we try to set it. If the device _does_ support it, the
user can configure the desired target temperature in the ViCare app.

I left the parameter in the template and only marked it as 'outdated'
- otherwise evcc errors on startup, if the parameter was previously
defined:
```
] FATAL [db:1] cannot create charger 'db:1': cannot create charger type 'template': invalid key: target_temperature
```
If we can instead mark it as 'hidden' or something that'd be better.

https://github.com/evcc-io/evcc/discussions/21850#discussioncomment-13604895