Update energinet-co2.yaml
The current API call to 'CO2EmisProg' does not include a 'start' parameter, so the response is limited to 100 records by default. Since the data is provided in 5-minute intervals and returned in descending order from midnight (local time), the API returns the last 100 forecast records furthest in the future, covering only the time range from 15:40 to 23:59.

As a result, evcc does not receive any forecast data for the earlier part of the day (00:00–15:40).

By adding 'start=now' to the API call, evcc receives the full forecast period.