Enable setting SEMP_BASE_URL for Home Assistant addon
When running as a Home Assistant addon, evcc is running in a container.
Therefore, the IP of the container is not helpful to create the hems URL
for SMA. It can already be overridden by an environment variable. This
change enables to expose this variable via Home Assistant so that it can
be set properly with the external URL.