Remove template for SOLARWATT Manager flex.
The template accessed an undocumented interface of the SOLARWATT Manager flex.
SOLARWATT explicitly warns against this and changes this interface without notice.
Now template no longer works. Adapting it appears to be time-consuming.

It is safer to access the connected devices directly.

The template will therefore be removed without replacement.

Please note: This is a breaking change for anyone who has managed to get the template working indirectly!