POC: add more resilience to tibber tariff ... in case of internet connect connection problems.
## Motivation
For the last couple of days my car wouldn't load simply because the planner wasn't able to retrieve current tariff data from Tibber. The cause was my internet provider doing some or other nightly maintenance, effectively shutting down my internet connection.

The thing, however, is that at that time evcc already had valid - albeit possibly outdated - tariff data available. It just wouldn't use it because of the internet connection error. Given the choice, I'd prefer to have a charged car, even at a higher cost than necessary, than to start the morning with an uncharged car.

So this patch to the Tibber tariff integration adds a fallback that will suppress the propagation of errors caused by network timeouts for a limited time (currently 36 hours - tbd) before propagating the problem to the caller.

No other functional changes to the code as far as I can see, so this should be a drop-in enhancement. It of course could be refined by using a tunable timeout in the template and possible other changes but the goal was to first provide a robust and resilient implementation.